Remove old main.js
