+++ models/group.model.js
@@ -1,0 +1,25 @@
+const mongoose = require('mongoose');
+
+const GroupSchema = new mongoose.Schema({
+  name: {
+    type: String,
+    required: true,
+    trim: true,
+    minlength: 3,
+    maxlength: 50
+  },
+  avatar: String,
+  members: [{ type: mongoose.Schema.Types.ObjectId, ref: 'User' }],
+  createdBy: { type: mongoose.Schema.Types.ObjectId, ref: 'User', required: true },
+  editHistory: [{
+    name: String,
+    avatar: String,
+    members: [mongoose.Schema.Types.ObjectId],
+    updatedAt: Date,
+    updatedBy: { type: mongoose.Schema.Types.ObjectId, ref: 'User' }
+  }],
+  createdAt: { type: Date, default: Date.now },
+  updatedAt: { type: Date, default: Date.now }
+});
+
+module.exports = mongoose.model('Group', GroupSchema);
