+++ package.json
@@ -19,8 +19,9 @@
     "helmet": "^7.0.0",
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^7.6.3",
-    "multer": "^1.4.5-lts.1",
+    "multer": "^1.4.5-lts.2",
     "socket.io": "^4.7.2",
-    "agenda": "^5.0.0"
+    "agenda": "^5.0.0",
+    "express-validator": "^7.0.1"
   }
 }
