+++ public/index.html
@@ -108,18 +108,6 @@
             <!-- Contacts List -->
             <div class="contacts-list" id="contacts-list" role="list">
                 <!-- Contacts will be dynamically added here -->
-                <div class="contact" role="listitem">
-                    <div class="contact-avatar">
-                        <img src="https://i.pravatar.cc/150?img=1" alt="Avatar de Jean">
-                        <span class="contact-status online"></span>
-                    </div>
-                    <div class="contact-info">
-                        <div class="contact-name">Jean Dupont</div>
-                        <div class="contact-last-msg">Salut, ça va ?</div>
-                    </div>
-                    <div class="contact-time">12:30</div>
-                    <div class="unread-count">3</div>
-                </div>
             </div>
         </aside>
         
@@ -158,42 +146,9 @@
                 </div>
             </header>
             
-            <!-- Messages Container -->
-            <section id="messages-container" class="messages-container" aria-live="polite">
-                <!-- Messages will be dynamically added here -->
-                <div class="date-separator">Aujourd'hui</div>
-                
-                <div class="message received" role="listitem">
-                    <div class="message-wrapper">
-                        <div class="message-content">
-                            Salut, comment ça va ?
-                        </div>
-                        <div class="message-info">
-                            <span class="timestamp">12:30</span>
-                        </div>
-                    </div>
-                </div>
-                
-                <div class="message sent" role="listitem">
-                    <div class="message-wrapper">
-                        <div class="message-content">
-                            Ça va bien merci ! Et toi ?
-                            <div class="message-actions">
-                                <button class="message-action-btn" aria-label="Répondre">
-                                    <i class="fas fa-reply"></i>
-                                </button>
-                                <button class="message-action-btn" aria-label="Réaction">
-                                    <i class="fas fa-smile"></i>
-                                </button>
-                            </div>
-                        </div>
-                        <div class="message-info">
-                            <span class="timestamp">12:32</span>
-                            <i class="fas fa-check read-receipt read"></i>
-                        </div>
-                    </div>
-                </div>
-            </section>
+              <section id="messages-container" class="messages-container" aria-live="polite">
+                  <!-- Messages will be dynamically added here -->
+              </section>
             
             <!-- Chat Input Area -->
             <footer class="chat-input-area">
@@ -300,6 +255,6 @@
     
     <!-- Socket.io -->
     <script src="/socket.io/socket.io.js"></script>
-    <script src="main.js" defer></script>
+    <script type="module" src="js/app.js"></script>
 </body>
 </html>
