+++ public/style.css
@@ -1013,3 +1013,33 @@ button {
     from { opacity: 1; transform: translateX(0); }
     to { opacity: 0; transform: translateX(100%); }
 }
+
+/* Skeleton Loading */
+.skeleton {
+    position: relative;
+    overflow: hidden;
+    background-color: var(--color-surface-light);
+    border-radius: 4px;
+}
+.skeleton::after {
+    content: "";
+    position: absolute;
+    top: 0;
+    left: -100%;
+    width: 100%;
+    height: 100%;
+    background: linear-gradient(90deg, transparent, rgba(255,255,255,0.1), transparent);
+    animation: skeleton-move 1.2s infinite;
+}
+@keyframes skeleton-move {
+    0% { left: -100%; }
+    100% { left: 100%; }
+}
+.contact-skeleton {
+    height: 40px;
+    margin: 0.5rem;
+}
+.message-skeleton {
+    height: 60px;
+    margin: 0.25rem;
+}
